# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Travelers API",
    "version": "1.0.0",
    "description": "API documentation for Travelers project"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 32
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 64
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "LoginUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "CreateStory": {
        "type": "object",
        "properties": {
          "img": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "maxLength": 150
          },
          "article": {
            "type": "string",
            "maxLength": 10000
          },
          "category": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        },
        "required": [
          "img",
          "title",
          "article",
          "category",
          "date"
        ],
        "additionalProperties": false
      },
      "UpdateStory": {
        "type": "object",
        "properties": {
          "img": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "maxLength": 150
          },
          "article": {
            "type": "string",
            "maxLength": 10000
          },
          "category": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdateUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 32
          },
          "description": {
            "type": "string",
            "maxLength": 150
          }
        },
        "additionalProperties": false
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout current user",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "User logged out"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh user session",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Session refreshed"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "Get categories list",
        "tags": [
          "Categories"
        ],
        "responses": {
          "200": {
            "description": "Categories list retrieved successfully"
          }
        }
      }
    },
    "/stories": {
      "get": {
        "summary": "Get all stories",
        "tags": [
          "Stories"
        ],
        "responses": {
          "200": {
            "description": "Stories list"
          }
        }
      },
      "post": {
        "summary": "Create story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Story created"
          }
        }
      }
    },
    "/stories/saved": {
      "get": {
        "summary": "Get saved stories",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Saved stories list"
          }
        }
      }
    },
    "/stories/my": {
      "get": {
        "summary": "Get my stories",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User stories list"
          }
        }
      }
    },
    "/stories/{storyId}": {
      "get": {
        "summary": "Get story by id",
        "tags": [
          "Stories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Story found"
          }
        }
      },
      "patch": {
        "summary": "Update story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Story updated"
          }
        }
      },
      "delete": {
        "summary": "Delete story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Story deleted"
          }
        }
      }
    },
    "/stories/{storyId}/save": {
      "post": {
        "summary": "Save story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Story saved"
          }
        }
      },
      "delete": {
        "summary": "Remove saved story",
        "tags": [
          "Stories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Story removed from saved"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get users list",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Users list retrieved successfully"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get current authenticated user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user data"
          }
        }
      }
    },
    "/users/me/avatar": {
      "patch": {
        "summary": "Update user avatar",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar updated successfully"
          }
        }
      }
    },
    "/users/me/update": {
      "patch": {
        "summary": "Update user profile",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "summary": "Get user by id",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          }
        }
      }
    }
  },
  "tags": []
}